Moved resolve_hostname and get_ip_from_resolved_hostname to os_net.c Moved unit test for that functions to test_os_net.c Minor changes in a debug message
