+++ src/error_messages/error_messages.h
@@ -286,10 +286,11 @@
 #define RL_REGEX_SYNTAX "(5107): Syntax error on tag '%s' in rule %d"
 
 /* Syslog output */
-#define XML_INV_CSYSLOG "(5301): Invalid client-syslog configuration."
+#define XML_INV_CSYSLOG    "(5301): Invalid client-syslog configuration."
+#define ERROR_SENDING_MSG  "(5302): Error sending message to '%s'."
 
 /* Integrator daemon */
-#define XML_INV_INTEGRATOR "(5302): Invalid integratord configuration."
+#define XML_INV_INTEGRATOR "(5310): Invalid integratord configuration."
 
 /* Agentless */
 #define XML_INV_AGENTLESS   "(7101): Invalid agentless configuration."
