+++ src/headers/os_utils.h
@@ -29,9 +29,5 @@ OSList *w_os_get_process_list();
 int w_is_file(const char * const file);
 /* Delete the process list */
 int w_del_plist(OSList *p_list);
-/* Resolve hostname */
-void resolve_hostname(char **hostname, int attempts);
-/* Extract IP from resolvedHostame */
-const char *get_ip_from_resolved_hostname(const char *resolved_hostname);
 
 #endif /* OS_UTILS_OP_H */
