+++ src/os_csyslogd/csyscom.c
@@ -75,7 +75,7 @@ void * csyscom_main(__attribute__((unused)) void * arg) {
 
     mdebug1("Local requests thread ready");
 
-    if (sock = OS_BindUnixDomain(DEFAULTDIR CSYS_LOCAL_SOCK, SOCK_STREAM, OS_MAXSTR), sock < 0) {
+    if (sock = OS_BindUnixDomain(CSYS_LOCAL_SOCK, SOCK_STREAM, OS_MAXSTR), sock < 0) {
         merror("Unable to bind to socket '%s': (%d) %s.", CSYS_LOCAL_SOCK, errno, strerror(errno));
         return NULL;
     }
