+++ src/shared/os_utils.c
@@ -10,7 +10,6 @@
 
 #include "shared.h"
 #include "os_utils.h"
-#include "os_net/os_net.h"
 
 #ifdef WIN32
 #include <tlhelp32.h>
@@ -118,7 +117,6 @@ OSList *w_os_get_process_list()
 }
 
 #endif
-
 /* Check if a file exists */
 int w_is_file(const char * const file)
 {
@@ -177,45 +175,6 @@ int w_del_plist(OSList *p_list)
     free(p_list);
 
     return (1);
-}
-
-void resolve_hostname(char **hostname, int attempts) {
-    char *tmp_str;
-    char *f_ip;
-
-    assert(hostname != NULL);
-    if (OS_IsValidIP(*hostname, NULL) == 1) {
-        return;
-    }
-
-    tmp_str = strchr(*hostname, '/');
-    if (tmp_str) {
-        *tmp_str = '\0';   // LCOV_EXCL_LINE
-    }
-
-    f_ip = OS_GetHost(*hostname, attempts);
-    if (f_ip) {
-        char ip_str[128] = {0};
-        snprintf(ip_str, 127, "%s/%s", *hostname, f_ip);
-        free(f_ip);
-        free(*hostname);
-        os_strdup(ip_str, *hostname);
-    } else {
-        char ip_str[128] = {0};
-        snprintf(ip_str, 127, "%s/", *hostname);
-        free(*hostname);
-        os_strdup(ip_str, *hostname);
-    }
-}
-
-const char *get_ip_from_resolved_hostname(const char *resolved_hostname){
-    char *tmp_str;
-    assert(resolved_hostname != NULL);
-
-    /* Check if we have a resolved_hostname or an IP */
-    tmp_str = strchr(resolved_hostname, '/');
-
-    return tmp_str ? ++tmp_str : resolved_hostname;
 }
 
 #ifdef WIN32
@@ -364,5 +323,4 @@ OSList *w_os_get_process_list()
     CloseHandle(hsnap);
     return (p_list);
 }
-
 #endif
